docs: add clarity in restrict fuzz amount to 2*96 wei (#62)
* docs: add clarity in restrict fuzz amount to 2*96 wei

Make it easier to see how to restrict the amount, reading fast is easy to miss the change from uint256 to uint96. Personally, I was looking into the `//snip comment` and trying to find a previous function or one it the cheatcodes  with that name. TLDR: Makes it easier to understand for noobies like me

* chore: ticks

Co-authored-by: Bjerg <onbjerg@users.noreply.github.com>